+++ packages/get-html/src/index.ts
@@ -1,0 +1,1 @@
+export { getHtml } from "./get-html";